Get rid of create state in Instance
We don't want to store any state in the Instance object, instead take
advantage of the new mechanism whereby the result of handle_create() is
passed to check_active().

This makes the autoscaling code uglier for now (since it must still store
the instance state in the instance), but this will go away when we move to
using co-routines to create resources in parallel.

Change-Id: Ib11c8da22900453eb9713eba5106859c8a09dbe2