Always use attributes from ResourceProxy
Prior to the creation of the StackDefinition class, when fetching resource
attributes in intrinsic functions we had to determine from the state of the
resource whether getting the attributes would be a valid thing to do, or if
we should just return None as a placeholder. This logic is now taken into
account when populating the ResourceProxy with attribute values, so
intrinsic functions can always rely on being able to call the
ResourceProxy's FnGetAtt() method.

This change allows us to provide more sophisticated placeholders in the
ResourceProxy when no attribute value is available.

Change-Id: I83d9c100f791dce8ce944de2c2bdaa8cbc0767f2
Story: #2003915
Task: 26802