Return placeholder values from FnGetAtt() pre-flight
When doing pre-flight validation, return a placeholder value from FnGetAtt
instead of None, if the type of the attribute is known from the attribute
schema.

This should allow some errors to be detected earlier. If the types are
mismatched this can be detected before creating any of the resources.

It should also allow some valid templates to pass validation where they
previously would have been rejected (e.g. when passing attributes to YAQL
functions that are not expecting None as an input).

Change-Id: Ib2479594625e4732e179984572666be7577228c0
Story: #2003915
Task: 26806