Make property values on delete more robust
When deleting a resource, we set its property values to the stored values
from the last update (since fc77e432978334687e143f0a8e8d69a14df2de1d) if we
know them.

This can return data that is not altogether meaningful in convergence when
there are no stored property values. Since we don't load every resource in
the stack in order to do the delete, we won't be able to get refids or
attributes from other resources. Any attributes will be replaced with None
based on special handling in the GetAtt function(s) that checks the status
of the target resource and gets back a completely bogus INIT_COMPLETE. We
wish to remove this special handling, but cannot do so without causing
errors.

Even when we have the stored property values, we can run into the same
issue because calling frozen_definition() has the side effect of also
trying the resolve e.g. metadata, even though we don't need that data.

Read only the frozen properties when they are available, and suppress
errors accessing the live properties when they are not. Since the
frozen_properties() decorator potentially suffers from the same issue (e.g.
during a signal), implement the fixes there and reuse it for the delete.

Change-Id: I6e23f326790e13a3fcb9eecfae8eb9327bbe33ba
Story: #2003915
Task: 26888