Split Properties.validate() method into two parts
Instead of passing a with_value argument to Properties.validate(), split it
into separate Properties.validate_template() and Properties.validate()
methods. The former validates only the structure of the template (no
unknown or unimplemented properties passed). The latter performs the deeper
validation that can only be done when values are supplied.

Change-Id: I331f9813fe8514f8d48125fed8b6f58b7ba2f47e