Validate placeholder values in properties
When doing pre-flight validation, only skip validating a property value
that depends on a resource if the value of the property resolves to None
(effectively this should only be resource attributes with no defined type).
Previously no property values that referenced a resource would be
validated pre-flight.

If a property value evaluates to a Placeholder, this will check that any
known constraints on the eventual value are consistent with the input
constraints of the property. If they are incompatible (for example, have
different custom constraint types), a validation error will be raised.

If a property evaluates to a list or dict (and not a Placeholder) it may
still contain an intrinsic function that currently evaluates either to None
or to a Placeholder. Those cases will, however, be handled by the
validation of the subproperties.

Change-Id: Ie377e6932e4a37ff57f4bb5ed2187e1013e4d12e